requirements.txt
flask==3.1.0
azure-storage-blob==12.24.0
azure-ai-textanalytics==5.3.0
